fix(sonarcloud): Descope since the same project is being used by all repos using the template
The same project is being used by all repos using the template.

To fix this the project needs to be created in advance and this is also a
pre-requisite as per https://github.com/SonarSource/sonarcloud-github-action#requirements.
This should be automated as part of the workflow. Otherwise, it becomes a bottleneck
for teams trying to onboard and waiting for admins to create the project.